Move GraphQL email content reader out of DSS
