+++ crates/complete_graph/src/lib.rs
@@ -14,8 +14,8 @@ mod sdl_test;
 pub use edges::{SoupEdges, SoupEmailThreadEdges};
 pub use graphql_common::GraphqlSoupRequestParts;
 pub use graphql_email::{
-    EmailContentKey, EmailContentLoad, EmailContentLoader, NoOpSoupEmailContentEdgeReader,
-    SoupEmailContentEdgeReader, email_content_loader,
+    EmailContentKey, EmailContentLoad, EmailContentLoader, EmailServiceEmailContentReader,
+    NoOpSoupEmailContentEdgeReader, SoupEmailContentEdgeReader, email_content_loader,
 };
 pub use graphql_notification::{
     EntityNotificationsLoader, SoupNotificationEdgeReader, entity_notifications_loader,
