+++ crates/graphql_email/Cargo.toml
@@ -8,6 +8,7 @@ publish = false
 async-graphql = { workspace = true, features = ["dataloader"] }
 chrono = { workspace = true }
 email = { path = "../email", default-features = false, features = ["ports"] }
+entity_access = { path = "../entity_access", default-features = false }
 macro_user_id = { path = "../macro_user_id" }
 tokio = { workspace = true }
 tracing = { workspace = true }
