+++ crates/graphql_email/src/lib.rs
@@ -6,7 +6,7 @@ mod loaders;
 mod objects;
 
 pub use loaders::{
-    EmailContentKey, EmailContentLoad, EmailContentLoader, NoOpSoupEmailContentEdgeReader,
-    SoupEmailContentEdgeReader, email_content_loader,
+    EmailContentKey, EmailContentLoad, EmailContentLoader, EmailServiceEmailContentReader,
+    NoOpSoupEmailContentEdgeReader, SoupEmailContentEdgeReader, email_content_loader,
 };
 pub use objects::{GraphqlSoupEmailMessage, load_latest_email_message};
