+++ crates/graphql_email/src/loaders/test.rs
@@ -1,15 +1,9 @@
-use std::{
-    collections::HashMap,
-    sync::{
-        Arc,
-        atomic::{AtomicUsize, Ordering},
-    },
+use std::sync::{
+    Arc,
+    atomic::{AtomicUsize, Ordering},
 };
 
-use email::domain::{
-    models::{EmailErr, ParsedMessage},
-    ports::EmailContentService,
-};
+use email::domain::models::EmailErr;
 use entity_access::domain::{
     models::{
         AccessLevel, BotId, CallChannelInfo, EntityAccessReceipt, EntityPermission, EntityType,
@@ -23,6 +17,24 @@ use macro_user_id::{
 };
 
 use super::*;
+
+#[derive(Clone, Default)]
+struct RecordingReader {
+    calls: Arc<AtomicUsize>,
+}
+
+impl SoupEmailContentEdgeReader for RecordingReader {
+    async fn get_email_content(
+        &self,
+        _user_id: &MacroUserIdStr<'static>,
+        keys: Vec<EmailContentKey>,
+    ) -> HashMap<EmailContentKey, EmailContentLoad> {
+        self.calls.fetch_add(1, Ordering::SeqCst);
+        keys.into_iter()
+            .map(|key| (key, EmailContentLoad::Missing))
+            .collect()
+    }
+}
 
 #[derive(Clone)]
 struct TestAccessService {
@@ -155,14 +167,50 @@ fn key(index: usize) -> EmailContentKey {
 }
 
 #[tokio::test]
+async fn batches_distinct_threads_in_one_reader_call() {
+    let reader = RecordingReader::default();
+    let user_id = MacroUserIdStr::try_from_email("reader@example.com").unwrap();
+    let loader = email_content_loader(user_id, reader.clone());
+    let first = key(1);
+    let second = key(2);
+
+    let loaded = loader.load_many(vec![first, second]).await.unwrap();
+
+    assert_eq!(reader.calls.load(Ordering::SeqCst), 1);
+    assert!(matches!(
+        loaded.get(&first),
+        Some(EmailContentLoad::Missing)
+    ));
+    assert!(matches!(
+        loaded.get(&second),
+        Some(EmailContentLoad::Missing)
+    ));
+}
+
+#[tokio::test]
+async fn rejects_oversized_batches_without_calling_the_reader() {
+    let reader = RecordingReader::default();
+    let user_id = MacroUserIdStr::try_from_email("reader@example.com").unwrap();
+    let loader = EmailContentLoader::new(user_id, reader.clone());
+    let keys = (0..=MAX_EMAIL_CONTENT_KEYS).map(key).collect::<Vec<_>>();
+
+    let error = loader.load(&keys).await.unwrap_err();
+
+    assert_eq!(reader.calls.load(Ordering::SeqCst), 0);
+    assert!(error.to_string().contains("at most 20 threads"));
+}
+
+#[tokio::test]
 async fn authorized_keys_reach_the_email_domain() {
     let content = Arc::new(RecordingContentService::default());
-    let service =
-        DssEmailContentReader::new(content.clone(), Arc::new(TestAccessService { allow: true }));
+    let reader = EmailServiceEmailContentReader::new(
+        content.clone(),
+        Arc::new(TestAccessService { allow: true }),
+    );
     let user_id = MacroUserIdStr::try_from_email("reader@example.com").unwrap();
     let requested = key(1);
 
-    let loaded = service.get_email_content(&user_id, vec![requested]).await;
+    let loaded = reader.get_email_content(&user_id, vec![requested]).await;
 
     assert_eq!(content.calls.load(Ordering::SeqCst), 1);
     assert!(matches!(
@@ -174,14 +222,14 @@ async fn authorized_keys_reach_the_email_domain() {
 #[tokio::test]
 async fn unauthorized_keys_do_not_reach_the_email_domain() {
     let content = Arc::new(RecordingContentService::default());
-    let service = DssEmailContentReader::new(
+    let reader = EmailServiceEmailContentReader::new(
         content.clone(),
         Arc::new(TestAccessService { allow: false }),
     );
     let user_id = MacroUserIdStr::try_from_email("reader@example.com").unwrap();
     let requested = key(1);
 
-    let loaded = service.get_email_content(&user_id, vec![requested]).await;
+    let loaded = reader.get_email_content(&user_id, vec![requested]).await;
 
     assert_eq!(content.calls.load(Ordering::SeqCst), 0);
     assert!(matches!(
