+++ services/document_storage_service/src/api/context.rs
@@ -147,7 +147,7 @@ pub(crate) type DssGraphqlSoupSchema = complete_graph::SharedSoupSchema<
     complete_graph::PropertiesEntityPropertyWriter<PropertiesService, EntityAccessService>,
     Arc<ai_tools::ToolNotificationService>,
     complete_graph::PropertiesEntityPropertyReader<PropertiesService, EntityAccessService>,
-    crate::api::graphql_soup::DssEmailContentReader<DssEmailService, EntityAccessService>,
+    complete_graph::EmailServiceEmailContentReader<DssEmailService, EntityAccessService>,
 >;
 
 type SystemPropertiesService = SystemPropertiesServiceImpl<PgSystemPropertiesRepository>;
