+++ services/document_storage_service/src/api/graphql_soup.rs
@@ -11,10 +11,6 @@ use axum_extra::extract::Cached;
 use complete_graph::GraphqlSoupRequestParts;
 use model_user::axum_extractor::OptionalMacroUserExtractor;
 
-mod email_content;
-
-pub(crate) use email_content::DssEmailContentReader;
-
 pub(crate) fn router() -> Router<ApiContext> {
     Router::new().route("/soup/graphql", get(graphiql).post(handler))
 }
@@ -67,7 +63,7 @@ async fn handler(State(state): State<ApiContext>, req: Request) -> Response {
         state.entity_access_service.clone(),
         macro_user_id.clone(),
     );
-    let email_content_reader = DssEmailContentReader::new(
+    let email_content_reader = complete_graph::EmailServiceEmailContentReader::new(
         state.soup_router_state.email_service(),
         state.entity_access_service.clone(),
     );
